main.py
import os
import openai
from PIL import Image, ImageDraw, ImageFont
import cv2
from googletrans import Translator
import requests

# Configuration
openai.api_key = 'YOUR_OPENAI_API_KEY'
deepseek_api_key = 'YOUR_DEEPSEEK_API_KEY'
jianyingpro_api_url = 'YOUR_JIANYINGPRO_API_URL'


def read_text_file(file_path):
    with open(file_path, 'r', encoding='utf-8') as file:
        return file.readlines()


def split_sentences(text_lines):
    sentences = []
    for line in text_lines:
        sentences.extend(line.strip().split('. '))
    return sentences


def ai_inference(prompt, use_chatgpt=True):
    if use_chatgpt:
        response = openai.Completion.create(
            engine="text-davinci-003",
            prompt=prompt,
            max_tokens=50
        )
        return response.choices[0].text.strip()
    else:
        response = requests.post(
            'https://api.deepseek.com/inference',
            headers={'Authorization': f'Bearer {deepseek_api_key}'},
            json={'prompt': prompt}
        )
        return response.json()['result']


def create_image_from_text(text):
    # Placeholder function to generate an image from text
    image = Image.new('RGB', (512, 512), color=(255, 255, 255))
    draw = ImageDraw.Draw(image)
    font = ImageFont.load_default()
    draw.text((10, 10), text, fill=(0, 0, 0), font=font)
    return image


def apply_random_motion_effect(image):
    # Placeholder function to add random motion effects to an image
    return image


def convert_to_hd(image):
    return image.resize((1920, 1080), Image.ANTIALIAS)


def save_image(image, file_path):
    image.save(file_path)


def import_to_jianyingpro(image_path):
    response = requests.post(
        jianyingpro_api_url,
        files={'file': open(image_path, 'rb')}
    )
    return response.json()


def translate_text(text, target_lang='en'):
    translator = Translator()
    return translator.translate(text, dest=target_lang).text


def batch_process(file_path, use_chatgpt=True):
    text_lines = read_text_file(file_path)
    sentences = split_sentences(text_lines)
    for sentence in sentences:
        prompt = translate_text(sentence)
        ai_result = ai_inference(prompt, use_chatgpt)
        image = create_image_from_text(ai_result)
        image = apply_random_motion_effect(image)
        image = convert_to_hd(image)
        image_path = os.path.join('output', f'{prompt[:10]}.png')
        save_image(image, image_path)
        import_to_jianyingpro(image_path)


if __name__ == '__main__':
    input_file = 'input/sample.txt'  # Example input file
    batch_process(input_file)