requirements.txt
openai
requests
Pillow
opencv-python
googletrans==4.0.0-rc1
pyinstaller